Move utils to rv module
