+++ src/detection/rv/od/commands/eval_model.py
@@ -6,7 +6,7 @@ import click
 
 from rv.od.commands.predict import _predict
 from rv.od.commands.eval_predictions import _eval_predictions
-from rv.od.commands.utils import (
+from rv.util import (
     download_if_needed, get_local_path, upload_if_needed, load_projects,
     make_temp_dir)
 from rv.od.commands.settings import planet_channel_order, temp_root_dir
