+++ src/detection/rv/od/commands/eval_predictions.py
@@ -7,7 +7,7 @@ import click
 
 from object_detection.utils import object_detection_evaluation, label_map_util
 
-from rv.od.commands.utils import (
+from rv.util import (
     download_if_needed, make_temp_dir, get_local_path, upload_if_needed,
     get_boxes_from_geojson, download_and_build_vrt)
 from rv.od.commands.settings import max_num_classes, temp_root_dir
