+++ src/detection/rv/od/commands/make_predict_chips.py
@@ -6,7 +6,7 @@ import click
 import numpy as np
 import rasterio
 
-from rv.od.commands.utils import load_window, save_img
+from rv.util import load_window, save_img
 from rv.od.commands.settings import planet_channel_order
