+++ src/detection/rv/od/commands/predict_on_chips.py
@@ -20,7 +20,7 @@ from object_detection.utils.np_box_list import BoxList
 
 from rv.od.commands.settings import (
     max_num_classes, min_score_threshold, line_thickness)
-from rv.od.commands.utils import save_img
+from rv.util import save_img
 
 image_size = (12, 8)
